# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server API e Websocket per il progetto HermesChat. Il server è stato sviluppato in Go con il framework Gin e utilizza MongoDB come database.",
        "title": "HermesChat API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.hermeschat.it:8090",
    "basePath": "/",
    "paths": {
        "/acceptFriend": {
            "post": {
                "description": "Accetta la richiesta di amicizia dell'utente loggato con l'utente passato come parametro",
                "produces": [
                    "application/json"
                ],
                "summary": "Accetta richiesta di amicizia",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Indice dell'utente da accettare",
                        "name": "idUser",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nickname dell'utente da accettare",
                        "name": "nickname",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nome dell'utente da accettare",
                        "name": "name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cognome dell'utente da accettare",
                        "name": "surname",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Immagine dell'utente da accettare",
                        "name": "image",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/blockUser": {
            "post": {
                "description": "Blocca l'utente passato come parametro dall'utente loggato",
                "produces": [
                    "application/json"
                ],
                "summary": "Blocca utente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username dell'utente da bloccare",
                        "name": "username",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/checkOtp": {
            "post": {
                "description": "Verifica che l'OTP inserito dall'utente sia lo stesso generato dal server e inviato via email",
                "produces": [
                    "application/json"
                ],
                "summary": "Verifica OTP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente da verificare",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OTP inserito dall'utente",
                        "name": "otp",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/declineFriend": {
            "post": {
                "description": "Rifiuta la richiesta di amicizia dell'utente loggato con l'utente passato come parametro",
                "produces": [
                    "application/json"
                ],
                "summary": "Rifiuta richiesta di amicizia",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Indice dell'utente da rifiutare",
                        "name": "idUser",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/getBlocked": {
            "post": {
                "description": "Restituisce tutti gli utenti bloccati dall'utente loggato",
                "produces": [
                    "application/json"
                ],
                "summary": "Ricerca utenti bloccati",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/getFriendRequests": {
            "post": {
                "description": "Restituisce tutti le richieste di amicizia dell'utente loggato",
                "produces": [
                    "application/json"
                ],
                "summary": "Ricerca richieste di amicizia utente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getFriends": {
            "post": {
                "description": "Restituisce tutti gli amici dell'utente loggato",
                "produces": [
                    "application/json"
                ],
                "summary": "Ricerca amici utente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getInfoUser": {
            "post": {
                "description": "Restituisce le informazioni dell'utente loggato",
                "produces": [
                    "application/json"
                ],
                "summary": "Informazioni utente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/getRequestSent": {
            "post": {
                "description": "Restituisce tutti le richieste di amicizia inviate dell'utente loggato",
                "produces": [
                    "application/json"
                ],
                "summary": "Ricerca richieste di amicizia inviate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/search": {
            "get": {
                "description": "Cerca fino a 100 utenti con username che inizia con la stringa passata come parametro",
                "produces": [
                    "application/json"
                ],
                "summary": "Ricerca utenti",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Caratteri con cui inizia l'username",
                        "name": "username",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/updateInfo": {
            "patch": {
                "description": "Aggiorna lo \"stato\" dell'utente loggato",
                "produces": [
                    "application/json"
                ],
                "summary": "Aggiorna informazioni utente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nuove informazioni da aggiornare",
                        "name": "newInfo",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}